+++ imposter.go
@@ -1,11 +1,20 @@
 package killgrave
 
-import "net/http"
+import (
+	"net/http"
+	"path"
+)
 
 // Imposter define an imposter structure
 type Imposter struct {
+	BasePath string
 	Request  Request  `json:"request"`
 	Response Response `json:"response"`
+}
+
+// CalculateFilePath calculate file path based on basePath of imposter directory
+func (i *Imposter) CalculateFilePath(filePath string) string {
+	return path.Join(i.BasePath, filePath)
 }
 
 // Request represent the structure of real request
@@ -18,8 +27,8 @@ type Request struct {
 
 // Response represent the structure of real response
 type Response struct {
-	Status      int     `json:"status"`
-	Body        string  `json:"body"`
-	BodyFile    *string `json:"bodyFile"`
-	Headers    *http.Header `json:"headers"`
+	Status   int          `json:"status"`
+	Body     string       `json:"body"`
+	BodyFile *string      `json:"bodyFile"`
+	Headers  *http.Header `json:"headers"`
 }