+++ server.go
@@ -40,6 +40,10 @@ func (s *Server) buildImposters() error {
 	files, _ := ioutil.ReadDir(s.impostersPath)
 
 	for _, f := range files {
+		if f.IsDir() {
+			continue
+		}
+
 		var imposter Imposter
 		if err := s.buildImposter(f.Name(), &imposter); err != nil {
 			return err
@@ -48,7 +52,9 @@ func (s *Server) buildImposters() error {
 		if imposter.Request.Endpoint == "" {
 			continue
 		}
-		s.router.HandleFunc(imposter.Request.Endpoint, ImposterHandler(imposter)).Methods(imposter.Request.Method)
+		s.router.HandleFunc(imposter.Request.Endpoint, ImposterHandler(imposter)).
+			Methods(imposter.Request.Method).
+			MatcherFunc(MatcherBySchema(imposter))
 	}
 
 	return nil
@@ -63,5 +69,7 @@ func (s *Server) buildImposter(imposterFileName string, imposter *Imposter) erro
 	if err := json.Unmarshal(bytes, imposter); err != nil {
 		return malformattedImposterError(fmt.Sprintf("error while unmarshall imposter file %s", f))
 	}
+	imposter.BasePath = s.impostersPath
+
 	return nil
 }