+++ CHANGELOG.md
@@ -14,3 +14,10 @@
 * Create an official docker image for the application
 * Update README.md with how to use the application with docker
 * Allow write headers for the response
+
+## v0.2.1 (2019/04/25)
+
+* Allow imposter's matching by request schema
+* Dynamic responses based on regex endpoint or request schema
+* Calculate files directory(body and schema) based on imposters path
+* Update REAMDE.md with resolved features and new future features